+++ crates/rpc-config/Cargo.toml
@@ -15,7 +15,7 @@ solders-account-decoder = { workspace = true }
 solders-pubkey = { workspace = true }
 solders-hash = { workspace = true }
 solders-signature = { workspace = true }
-solders-transaction-status = { workspace = true }
+solders-transaction-status-enums = { workspace = true }
 solders-rpc-filter = { workspace = true }
 solders-commitment-config = { workspace = true }
 solana-sdk = { workspace = true }
