+++ crates/transaction-status-enums/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "solders-transaction-status-enums"
+version = "0.14.4"
+edition = "2021"
+include = ["/src"]
+license = "MIT"
+repository = "https://github.com/kevinheavey/solders"
+
+[dependencies]
+serde = { workspace = true }
+pyo3 = { workspace = true }
