+++ crates/transaction-status-enums/src/lib.rs
@@ -1,0 +1,37 @@
+use pyo3::prelude::*;
+use serde::{Deserialize, Serialize};
+
+/// Levels of transaction detail to return in RPC requests.
+#[derive(Debug, Clone, Copy, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[pyclass(module = "solders.transaction_status")]
+pub enum TransactionDetails {
+    Full,
+    Signatures,
+    #[serde(rename = "none")]
+    None_,
+}
+
+impl Default for TransactionDetails {
+    fn default() -> Self {
+        Self::Full
+    }
+}
+
+/// Encoding options for transaction data.
+#[pyclass(module = "solders.transaction_status")]
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum UiTransactionEncoding {
+    Binary, // Legacy. Retained for RPC backwards compatibility
+    Base64,
+    Base58,
+    Json,
+    JsonParsed,
+}
+
+impl Default for UiTransactionEncoding {
+    fn default() -> Self {
+        Self::Base64
+    }
+}
