+++ crates/transaction-status/Cargo.toml
@@ -24,5 +24,6 @@ solders-hash = { workspace = true }
 solders-message = { workspace = true }
 solders-transaction = { workspace = true }
 solders-transaction-error = { workspace = true }
+solders-transaction-status-enums = { workspace = true }
 solders-commitment-config ={ workspace = true }
 bincode ={ workspace = true }
