+++ crates/transaction-status/src/lib.rs
@@ -15,6 +15,8 @@ use solders_transaction_error::{
     TransactionErrorInstructionError, TransactionErrorInsufficientFundsForRent,
     TransactionErrorType,
 };
+use solders_transaction_status_enums::{TransactionDetails, UiTransactionEncoding};
+
 use std::str::FromStr;
 pub mod tmp_transaction_status;
 
@@ -51,41 +53,6 @@ use tmp_transaction_status::{
     UiTransactionReturnData, UiTransactionStatusMeta as UiTransactionStatusMetaOriginal,
     UiTransactionTokenBalance as UiTransactionTokenBalanceOriginal,
 };
-
-/// Encoding options for transaction data.
-#[pyclass(module = "solders.transaction_status")]
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum UiTransactionEncoding {
-    Binary, // Legacy. Retained for RPC backwards compatibility
-    Base64,
-    Base58,
-    Json,
-    JsonParsed,
-}
-
-impl Default for UiTransactionEncoding {
-    fn default() -> Self {
-        Self::Base64
-    }
-}
-
-/// Levels of transaction detail to return in RPC requests.
-#[derive(Debug, Clone, Copy, Eq, Hash, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-#[pyclass(module = "solders.transaction_status")]
-pub enum TransactionDetails {
-    Full,
-    Signatures,
-    #[serde(rename = "none")]
-    None_,
-}
-
-impl Default for TransactionDetails {
-    fn default() -> Self {
-        Self::Full
-    }
-}
 
 #[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
 #[serde(rename_all = "camelCase")]
